pytest: Separate blob tests
